requirements.txt
pandas==2.2.0
tabula_py==1.4.2
